remove errors.h dependency on helpers.h, add SIR_PERF_PROFILE (#255)
* Adds SIR_PERF_PROFILE
