+++ sirplatform.mk
@@ -124,7 +124,7 @@ ifeq ($(SUNPRO),1)
   MMDOPT=-xMMD
   PTHOPT=-mt=yes
 else
-  FORTIFY_FLAGS=-D_FORTIFY_SOURCE=2
+  FORTIFY_FLAGS?=-D_FORTIFY_SOURCE=2
   MMDOPT=-MMD
   PTHOPT=-pthread
 endif
